don't rate limit key when registering handlers
Problem:
EnqueueKey rate limits the key and is called by sharedcontroller's RegisterHandler.
This is an issue because on startup we start enqueuing each key for each handler, causing the rate limit duration spike to hours very quickly.

Solution:
Adding the key to workqueue directly in EnqueueKey. This is what 2.4.x did as well,
https://github.com/rancher/norman/blob/release/v2.4/controller/generic_controller.go#L187