Add option to only enqueue if actually updated
This option causes the handle function to only proceed if the
object was actually updated. This is intended to be used by
applications with many objects and/or controllers types that
have alternative means of rerunning when necessary. When the
informer's resync their cache the update function is run.
This reducing the overhead but has the tradeoff of not
rerunning handlers that may rely on external objects or
services or experience a bug causing them to not retry even
if it was necessary. The result is that running the handler
func on resync will mostly only serve the purpose of catching
missed cache events.