Transition to simpler key management (#80)
Move away from key hierarchy keys, into a simpler data slicing instead.

The use of key hierarchy, saving the DEKs alongside the data, is very
effective when there is a need for long-term storage. As Lasso's KEK is
not persisted externally, once the instance restarts the encrypted data
can no longer be used as the DEKs can't be decrypted.

Due to that, a more performant approach would be do away with the KEK, and
keep all the DEKs in memory. The encrypted data now has a keyID, as opposed
to having an encrypted DEK and its nonce.

Signed-off-by: Paulo Gomes <pjbgf@linux.com>
Signed-off-by: Silvio Moioli <silvio@moioli.net>
Co-authored-by: Silvio Moioli <silvio@moioli.net>