Drop the attachdriver mechanism
After security assessment it was determined storing filterable/sortable
columns on disk is acceptable, as none of the information is expected
to be sensible.

This removes the attachdriver mechanism that created one DB on the disk
and another, optionally in memory, as attached DB.

Only one DB is created now.

This change was driven by the need of dropping the shared cache SQLite
mode:

https://sqlite.org/sharedcache.html

which is in turn needed in order to activate WAL mode, which allows
for better performance and less locking (including at CREATE INDEX
time).

Signed-off-by: Silvio Moioli <silvio@moioli.net>